# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beyond Auth",
    "description": "Per-project authentication and authorization service.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/v1/orgs": {
      "get": {
        "tags": [
          "orgs"
        ],
        "summary": "List all orgs the authenticated user is a member of, cursor-paginated.",
        "operationId": "list_orgs",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "orgs"
        ],
        "summary": "Create a new organization. The authenticated user becomes the owner.\nReturns 409 if the slug is already taken.",
        "operationId": "create_org",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register a new user with email and password. Creates the user, a personal org, and\na session in a single transaction. Returns 409 if the email is already registered.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Password validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Return the authenticated user's profile, primary email, and personal org.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Soft-delete the authenticated user and their personal org. All active sessions are\ninvalidated immediately. This action is permanent — deleted accounts cannot be recovered.",
        "operationId": "delete_me",
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update the authenticated user's profile. Only fields present in the request body are changed.",
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "description": "Returned on successful login or signup. Contains the new session alongside the user,\nprimary email, and personal org context.",
        "required": [
          "user",
          "email",
          "org",
          "session"
        ],
        "properties": {
          "email": {
            "$ref": "#/components/schemas/EmailBody"
          },
          "org": {
            "$ref": "#/components/schemas/OrgBody"
          },
          "session": {
            "$ref": "#/components/schemas/SessionBody"
          },
          "user": {
            "$ref": "#/components/schemas/UserBody"
          }
        }
      },
...
      "CreateOrgRequest": {
        "type": "object",
        "description": "Request to create a new organization.",
        "required": [
          "name",
          "metadata"
        ],
        "properties": {
          "metadata": {
            "type": [
              "object",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL-safe identifier. Defaults to a slugified version of `name` if omitted."
          }
        }
      },
...
      "EmailBody": {
        "type": "object",
        "description": "Primary email address for the user.",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Null if the address has not been verified."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Wire-format error body returned on all non-2xx responses.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code, e.g. `\"invalid_credentials\"`."
          },
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional actionable guidance present on configuration-gate errors."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Top-level error envelope for all API error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "description": "Current authenticated user context.",
        "required": [
          "user",
          "email",
          "org"
        ],
        "properties": {
          "email": {
            "$ref": "#/components/schemas/EmailBody"
          },
          "org": {
            "$ref": "#/components/schemas/OrgBody",
            "description": "The user's personal org. Multi-org memberships are listed via `GET /v1/orgs`."
          },
          "user": {
            "$ref": "#/components/schemas/UserBody"
          }
        }
      },
...
      "OrgBody": {
        "type": "object",
        "description": "The user's personal org.",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier, unique across all orgs."
          }
        }
      },
      "OrgResponse": {
        "type": "object",
        "description": "An organization the authenticated user is a member of.",
        "required": [
          "id",
          "name",
          "slug",
          "metadata",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary JSON metadata."
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier, unique across all orgs."
          }
        }
      },
      "OrgsResponse": {
        "type": "object",
        "description": "Cursor-paginated list of orgs.",
        "required": [
          "orgs",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "next_page": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor — pass as `after` to retrieve the next page."
          },
          "orgs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgResponse"
            }
          }
        }
      },
...
      "SessionBody": {
        "type": "object",
        "description": "Session credential returned on login or signup.",
        "required": [
          "id",
          "token",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "description": "Opaque bearer token — store securely, transmit as `Authorization: Bearer <token>`."
          }
        }
      },
...
      "SignupRequest": {
        "type": "object",
        "description": "Registration request. Creates a user, a personal org, and a session in one call.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name for the user's personal org. Defaults to the local part of the email."
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UpdateMeRequest": {
        "type": "object",
        "description": "Partial update for the user's personal org. All fields are optional; omitted fields\nare left unchanged.",
        "required": [
          "metadata"
        ],
        "properties": {
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ],
            "description": "Arbitrary JSON merged into the org's metadata field (full replacement, not merge)."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name for the user and their personal org."
          },
          "slug": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL-safe org identifier. Must be unique. Returns 409 if already taken."
          }
        }
      },
...
      "UserBody": {
        "type": "object",
        "description": "Core user fields.",
        "required": [
          "id",
          "primary_org_id",
          "name",
          "metadata",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary JSON metadata stored on the personal org."
          },
          "name": {
            "type": "string",
            "description": "Display name (mirrors the personal org name)."
          },
          "primary_org_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user's personal org. Always present; created at signup."
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_<id>_<secret>"
      }
    }
  },
...
}